refactor RIP
