+++ lib/models/app_store.dart
@@ -1,0 +1,3 @@
+class AppState {
+  
+}
