+++ lib/reducers/app_state_reducer.dart
@@ -1,0 +1,5 @@
+import 'package:my_project/models/app_store.dart';
+
+AppState appStateReducer(AppState state, action) {
+  return AppState();
+}
