+++ lib/routes.dart
@@ -1,0 +1,3 @@
+class FxbMarketDataRoutes {
+  static final home = '/';
+}
